+++ crates/infrastructure/tests/test_cache_database_postgres.rs
@@ -503,6 +503,7 @@ mod serial_tests {
         let mut account = AccountAny::Cash(CashAccount::new(
             cash_account_state_million_usd("1000000 USD", "0 USD", "1000000 USD"),
             false,
+            false,
         ));
         let last_event = account.last_event().unwrap();
         if last_event.base_currency.is_some() {